+++ CHANGELOG.md
@@ -81,7 +81,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed raw sql in access to be compatible with postgresql
 - Fixed copy bug in import script -> Copied to wrong location, when path had tailing '/'
 - PipeStart: Empty labels in annotask not possible anymore
-- SIA: Endless image loading bug
+- SIA: 
+  - Endless image loading bug
+  - Filter Bug -> Do not copy annotations from previous image when filter is active!
 
 ## Removed
 - removed lost-cv images. This is now integrated into the lost image